+++ src/server.rs
@@ -24,11 +24,10 @@ use bevy::{
 use crate::core::{
     channels::{ReplicationChannel, RepliconChannels},
     common_conditions::{server_just_stopped, server_running},
-    connected_clients::{
-        client_visibility::Visibility, ClientBuffers, ConnectedClient, ConnectedClients,
-        VisibilityPolicy,
-    },
     ctx::SerializeCtx,
+    replicated_clients::{
+        client_visibility::Visibility, ClientBuffers, ReplicatedClients, VisibilityPolicy,
+    },
     replication_registry::ReplicationRegistry,
     replication_rules::ReplicationRules,
     replicon_server::RepliconServer,
@@ -75,7 +74,7 @@ impl Plugin for ServerPlugin {
             .init_resource::<ServerTick>()
             .init_resource::<ClientBuffers>()
             .init_resource::<ClientEntityMap>()
-            .insert_resource(ConnectedClients::new(self.visibility_policy))
+            .insert_resource(ReplicatedClients::new(self.visibility_policy))
             .add_event::<ServerEvent>()
             .configure_sets(
                 PreUpdate,
@@ -157,7 +156,7 @@ impl ServerPlugin {
     fn handle_connections(
         mut server_events: EventReader<ServerEvent>,
         mut entity_map: ResMut<ClientEntityMap>,
-        mut connected_clients: ResMut<ConnectedClients>,
+        mut replicated_clients: ResMut<ReplicatedClients>,
         mut server: ResMut<RepliconServer>,
         mut client_buffers: ResMut<ClientBuffers>,
     ) {
@@ -165,11 +164,11 @@ impl ServerPlugin {
             match *event {
                 ServerEvent::ClientDisconnected { client_id, .. } => {
                     entity_map.0.remove(&client_id);
-                    connected_clients.remove(&mut client_buffers, client_id);
+                    replicated_clients.remove(&mut client_buffers, client_id);
                     server.remove_client(client_id);
                 }
                 ServerEvent::ClientConnected { client_id } => {
-                    connected_clients.add(&mut client_buffers, client_id);
+                    replicated_clients.add(&mut client_buffers, client_id);
                 }
             }
         }
@@ -177,12 +176,12 @@ impl ServerPlugin {
 
     fn cleanup_acks(
         update_timeout: Duration,
-    ) -> impl FnMut(ResMut<ConnectedClients>, ResMut<ClientBuffers>, Res<Time>) {
-        move |mut connected_clients: ResMut<ConnectedClients>,
+    ) -> impl FnMut(ResMut<ReplicatedClients>, ResMut<ClientBuffers>, Res<Time>) {
+        move |mut replicated_clients: ResMut<ReplicatedClients>,
               mut client_buffers: ResMut<ClientBuffers>,
               time: Res<Time>| {
             let min_timestamp = time.elapsed().saturating_sub(update_timeout);
-            for client in connected_clients.iter_mut() {
+            for client in replicated_clients.iter_mut() {
                 client.remove_older_updates(&mut client_buffers, min_timestamp);
             }
         }
@@ -191,7 +190,7 @@ impl ServerPlugin {
     fn receive_acks(
         change_tick: SystemChangeTick,
         mut server: ResMut<RepliconServer>,
-        mut connected_clients: ResMut<ConnectedClients>,
+        mut replicated_clients: ResMut<ReplicatedClients>,
         mut client_buffers: ResMut<ClientBuffers>,
     ) {
         for (client_id, message) in server.receive(ReplicationChannel::Init) {
@@ -200,7 +199,7 @@ impl ServerPlugin {
             while cursor.position() < message_end {
                 match bincode::deserialize_from(&mut cursor) {
                     Ok(update_index) => {
-                        let client = connected_clients.client_mut(client_id);
+                        let client = replicated_clients.client_mut(client_id);
                         client.acknowledge(
                             &mut client_buffers,
                             change_tick.this_run(),
@@ -221,7 +220,7 @@ impl ServerPlugin {
         change_tick: SystemChangeTick,
         mut set: ParamSet<(
             &World,
-            ResMut<ConnectedClients>,
+            ResMut<ReplicatedClients>,
             ResMut<ClientEntityMap>,
             ResMut<DespawnBuffer>,
             ResMut<RemovalBuffer>,
@@ -235,8 +234,8 @@ impl ServerPlugin {
     ) -> bincode::Result<()> {
         replicated_archetypes.update(set.p0(), &rules);
 
-        let connected_clients = mem::take(&mut *set.p1()); // Take ownership to avoid borrowing issues.
-        messages.prepare(connected_clients);
+        let replicated_clients = mem::take(&mut *set.p1()); // Take ownership to avoid borrowing issues.
+        messages.prepare(replicated_clients);
 
         collect_mappings(&mut messages, &mut set.p2())?;
         collect_despawns(&mut messages, &mut set.p3())?;
@@ -253,7 +252,7 @@ impl ServerPlugin {
         entities_with_removals.clear();
 
         let mut client_buffers = mem::take(&mut *set.p5());
-        let connected_clients = messages.send(
+        let replicated_clients = messages.send(
             &mut set.p6(),
             &mut client_buffers,
             **server_tick,
@@ -262,7 +261,7 @@ impl ServerPlugin {
         )?;
 
         // Return borrowed data back.
-        *set.p1() = connected_clients;
+        *set.p1() = replicated_clients;
         *set.p5() = client_buffers;
 
         Ok(())
@@ -271,12 +270,12 @@ impl ServerPlugin {
     fn reset(
         mut server_tick: ResMut<ServerTick>,
         mut entity_map: ResMut<ClientEntityMap>,
-        mut connected_clients: ResMut<ConnectedClients>,
+        mut replicated_clients: ResMut<ReplicatedClients>,
         mut client_buffers: ResMut<ClientBuffers>,
     ) {
         *server_tick = Default::default();
         entity_map.0.clear();
-        connected_clients.clear(&mut client_buffers);
+        replicated_clients.clear(&mut client_buffers);
     }
 }