+++ CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `ServerEventAppExt::make_independent` to let events be triggered without waiting for replication on the same tick.
 
+### Changed
+
+- Rename `connected_clients` into `replicated_clients`.
+- Rename `ConnectedClients` to `ReplicatedClients`.
+- Rename `ConnectedClient` to `ReplicatedClient`.
+
 ## [0.27.0] - 2024-07-04
 
 ### Changed